Folder renaming to avoid use numbers to sort them.
