+++ cpp-adapter-pattern/main.cpp
