+++ cpp-macro-compilation/main.cpp
