+++ cpp-relationship-between-classes/aggregation.cpp
