+++ cpp-signals/main.cpp
