+++ cpp-templates-with-interfaces/main.cpp
