+++ cpp-templates/main.cpp
