+++ cpp-threads-with-shared-memory-and-callbacks/api/IMonitorApi.h
