+++ cpp-threads-with-shared-memory-and-callbacks/include/Monitor.h
