+++ cpp-threads-with-shared-memory-and-callbacks/interfaces/IPerson.h
