+++ cpp-threads-with-shared-memory-and-callbacks/src/Monitor.cpp
