+++ goatnative-inject-investigation/Injector.h
