+++ goatnative-inject-investigation/main.cpp
