+++ handler-factory/src/XMLParser.cpp
